[F, #2] New file
Avoiding long files
MV LogUploader -> Upload Confs.
